requirements.txt
asgiref==3.8.1
bcc==0.1.10
certifi==2025.1.31
charset-normalizer==3.4.1
coreapi==2.3.3
coreschema==0.0.4
Django==5.1.7
django-cors-headers==4.7.0
django-extensions==3.2.3
djangorestframework==3.15.2
idna==3.10
itypes==1.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.2.3
requests==2.32.3
setuptools==76.0.0
sqlparse==0.5.3
traitlets==5.14.3
traittypes==0.2.1
uritemplate==4.1.1
urllib3==2.3.0
